backend/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gTTS==2.5.0
python-dotenv==1.0.1
# For production storage
# django-storages==1.14.2
# boto3==1.34.39 